requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1
xgboost==2.0.3
lightgbm==4.3.0
python-pptx==0.6.23 